# Conflicts:
#	src/main/java/com/pallow/pallow/domain/user/entity/User.java

+++ src/main/java/com/pallow/pallow/domain/profile/service/ProfileService.java
@@ -27,7 +27,6 @@ public class ProfileService {
     public ProfileResponseDto getProfile(Long userId) {
         Profile foundUser = profileRepository.findById(userId)
                 .orElseThrow(() -> new CustomException(ErrorType.NOT_FOUND_USER));
-
         return new ProfileResponseDto(foundUser);
     }
 
@@ -42,12 +41,23 @@ public class ProfileService {
     public ProfileResponseDto updateProfile(Long userId, ProfileRequestDto requestDto, User user) {
         Profile foundUser = profileRepository.findById(userId)
                 .orElseThrow(() -> new CustomException(ErrorType.NOT_FOUND_USER));
+        if (!isSameIdAndUser(userId, user)) {
+            throw new CustomException(ErrorType.USER_MISMATCH_ID);
+        }
         foundUser.update(requestDto);
         return new ProfileResponseDto(foundUser);
     }
 
     @Transactional
     public void deleteProfile(Long userId, User user) {
+        if (!isSameIdAndUser(userId, user)) {
+            throw new CustomException(ErrorType.USER_MISMATCH_ID);
+        }
         profileRepository.deleteById(userId);
     }
+
+    private boolean isSameIdAndUser(Long userId, User user) {
+        return user.getId().equals(userId);
+    }
+
 }
